--- .env.example
@@ -1,2 +1,0 @@
-TOKEN=EXAMPLETOKEN1234
-NASA=EXAMPLEXYZABC
