--- dates.js
@@ -1,69 +1,0 @@
-const autoGenDates = [
-    "2_1*C 6/7",
-    "2_2*A 1/2",
-    "2_3*B 1/2",
-    "2_4*C 1/2",
-    "2_5*A 4/5",
-    "2_8*B 4/5",
-    "2_9*C 4/5",
-    "2_10*A 6/7",
-    "2_11*B 6/7",
-    "2_12*C 6/7",
-    "2_22*A 1/2",
-    "2_23*B 1/2",
-    "2_24*C 1/2",
-    "2_25*A 4/5",
-    "2_26*B 4/5",
-    "2_29*C 4/5",
-    "2_30*A 6/7",
-    "2_31*B 6/7",
-    "3_5*C 6/7",
-    "3_6*A 1/2",
-    "3_7*B 1/2",
-    "3_8*C 1/2",
-    "3_9*A 4/5",
-    "3_12*B 4/5",
-    "3_13*C 4/5",
-    "3_14*A 6/7",
-    "3_15*B 6/7",
-    "3_16*C 6/7",
-    "3_19*A 1/2",
-    "3_20*B 1/2",
-    "3_21*C 1/2",
-    "3_22*A 4/5",
-    "3_23*B 4/5",
-    "3_26*C 4/5",
-    "3_27*A 6/7",
-    "3_28*B 6/7",
-    "3_29*C 6/7",
-    "3_30*A 1/2",
-    "4_3*B 1/2",
-    "4_4*C 1/2",
-    "4_5*A 4/5",
-    "4_6*B 4/5",
-    "4_7*C 4/5",
-    "4_10*A 6/7",
-    "4_11*B 6/7",
-    "4_12*C 6/7",
-    "4_13*A 1/2",
-    "4_14*B 1/2",
-    "4_17*C 1/2",
-    "4_18*A 4/5",
-    "4_19*B 4/5",
-    "4_20*C 4/5",
-    "4_21*A 6/7",
-    "4_24*B 6/7",
-    "4_25*C 6/7",
-    "4_26*A 1/2",
-    "4_27*B 1/2",
-    "4_28*C 1/2",
-    "5_1*A 4/5",
-    "5_2*B 4/5",
-    "5_3*C 4/5",
-    "5_4*A 6/7",
-    "5_7*B 6/7",
-    "5_8*C 6/7",
-    "5_9*A 1/2"
-];
-
-module.exports = autoGenDates;
